Volume Fixes
Fixed starting volume to not be at maximum.
Created a new volume slider for sound effects, which plays a sample sound effect when the player lets the slider go.
Changed it so that the explosion sound is now produced by the drones instead of the player.
The explosion that happens if a player dies to a drone is now audible.