Various Fixes + additions
Added text to the game over screen which explains the reason for the game over, as well as text that says the player's final score.
Made tree text a child object of trees and fixed the text alignment.
Made it so that other power ups can be picked up while you have a power up active, turning the active one off.
Simplified the code for enemy and tree movement.
Updated the bounds for object movements.
Added player prefs to store volume preferences and set these at the start of each game. Also added a functioning high score system with player prefs.
Made it so that drones give points when destroyed (except when the player collides with them without armour).
Fixed an issue in which one stick would count as two when being fired at trees or drones.
Added a new powerup that allows for firing additional sticks left, right and behind.
Made it so that tree text is bigger when it reaches a certain point without the player having completed it.
Adjusted pause code to be cleaner.