New input system and object pooling
Integrated the new input system.
Normalized player movement.
Updated the bounds for tree, power up and drone movement.
Added object pooling for sticks, drones, trees, sound effects and explosions.
General cleaning up of code.
There is a known issue in which the same object is returned to the pool twice.